refactor(core): aplicar Arquitectura Limpia y principios DRY en Casos de Uso
Este refactor mayor estandariza cómo se aplican las reglas de seguridad y de negocio en los módulos de Gym, Routine y Exercise:

1. Capa de Aplicación: Se extrajo la lógica repetida de autorización y aislamiento por gimnasio hacia `SecurityUtils` y `RoutineAccessValidator`. Todos los casos de uso son ahora más pequeños, limpios y enfocados únicamente en la orquestación.
2. Capa de Dominio: Se eliminó la validación de roles de la entidad `Routine`. El dominio ahora impone estrictamente las reglas de estado (el "qué"), mientras que la capa de aplicación impone las reglas de seguridad (el "quién").
3. Seguridad: Se previnieron vulnerabilidades IDOR validando explícitamente el `gymId` en todas las creaciones, actualizaciones y lecturas de recursos del sistema.

+++ src/main/java/com/trainingapp/trainingapp/infrastructure/repository/jpa/repository/user/UserJpaRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface UserJpaRepository extends JpaRepository<UserJpaEntity, Integer> {
+public interface UserJpaRepository extends JpaRepository<UserJpaEntity, Long> {
 
     Optional<UserJpaEntity> findByEmail(String email);
 }
